Updated Landing and added Footer
* Updated landing text and made cards link to respective pages

*Added Footer to all pages
