+++ src/app/page.tsx
@@ -1,6 +1,7 @@
 "use client";
+import Footer from "@/components/Footer";
 import Hero from "@/components/landing/Hero";
-import LandingAbout from "@/components/landing/LandingAbout";
+import Landing from "@/components/landing/Landing";
 
 const Website = () => {
   return (
@@ -8,7 +9,8 @@ const Website = () => {
       {/* Applied a noise bg to the page for added depth */}
       <div className="noise-bg" />
       <Hero />
-      <LandingAbout />
+      <Landing />
+      <Footer />
     </div>
   );
 };
