+++ src/app/projects/page.tsx
@@ -2,11 +2,13 @@
 
 import React from "react";
 import ProjectsPage from "@/components/projects/ProjectsPage";
+import Footer from "@/components/Footer";
 
 export default function Projects() {
   return (
     <div>
       <ProjectsPage />
+      <Footer />
     </div>
   );
 }
